pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]>=7"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
git_describe_command= 'git describe --tags --abbrev=0 --dirty --match "sane_controlnet_aux_v*"'
tag_regex = "^sane_controlnet_aux_v(?P<version>.*)$"

[project]
name = "sane-controlnet-aux"
dynamic = ["version"]
description = "A fork of ComfyUI's ControlNet auxiliary preprocessors for use outside of ComfyUI."
readme = "README.md"
authors = [{ name = "Vedat Baday <badayvedat@gmail.com>" }]
requires-python = ">=3.8"
dependencies = [
    "torch",
    "huggingface_hub",
    "scipy",
    "opencv-python-headless>=4.7.0.72",
    "numpy",
    "Pillow",
    "torchvision",
    "scikit-image",
    "timm==0.6.13",
]

[project.optional-dependencies]
test = [
    "pytest<8",
    "pytest-xdist",
]
dev = [
    "sane-controlnet-aux[test]",
]

[project.urls]
repository = "https://github.com/badayvedat/sane-controlnet-aux"


[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
